refactor: modularize agent modal and dashboard
