+++ apps/web/package.json
@@ -20,7 +20,8 @@
     "react-dom": "18.3.1",
     "recharts": "^2.15.4",
     "tailwind-merge": "^3.3.1",
-    "tailwindcss": "^3.4.10"
+    "tailwindcss": "^3.4.10",
+    "swr": "^2.3.0"
   },
   "devDependencies": {
     "@playwright/test": "^1.56.0",
