+++ apps/web/src/app/api/agents/[id]/files/[fileId]/route.ts
@@ -1,4 +1,6 @@
-import { forwardToEnacom } from '../../../_utils'
+import { forwardToEnacom } from '@/lib/api/forwardToEnacom'
+
+export const dynamic = 'force-dynamic'
 
 export async function DELETE(_req: Request, { params }: { params: { id: string; fileId: string } }) {
   return forwardToEnacom(`/agents/${params.id}/files/${params.fileId}`, { method: 'DELETE' })
