+++ apps/web/src/app/api/agents/[id]/metrics/route.ts
@@ -1,4 +1,6 @@
-import { forwardToEnacom } from '../../_utils'
+import { forwardToEnacom } from '@/lib/api/forwardToEnacom'
+
+export const dynamic = 'force-dynamic'
 
 export async function GET(
   _request: Request,
