+++ apps/web/src/app/api/agents/[id]/traces/route.ts
@@ -1,5 +1,7 @@
 import { NextRequest } from 'next/server'
-import { forwardToEnacom } from '../../_utils'
+import { forwardToEnacom } from '@/lib/api/forwardToEnacom'
+
+export const dynamic = 'force-dynamic'
 
 export async function GET(req: NextRequest, { params }: { params: { id: string } }) {
   const take = req.nextUrl.searchParams.get('take')
