+++ apps/web/src/app/api/dashboard/leaderboard/route.ts
@@ -1,5 +1,7 @@
-import { forwardToEnacom } from '../../agents/_utils';
+import { forwardToEnacom } from '@/lib/api/forwardToEnacom'
+
+export const dynamic = 'force-dynamic'
 
 export async function GET() {
-  return forwardToEnacom('/dashboard/leaderboard');
+  return forwardToEnacom('/dashboard/leaderboard')
 }
